+++ novawallet/Common/Extension/UIKit/Style/IconDetailsView+Style.swift
